Add JUnit test (chat event)
