+++ src/test/skript/junit/ExprMessage.sk
@@ -1,0 +1,19 @@
+test "ExprMessageJUnit" when running JUnit:
+	set {_tests::1} to "set message"
+	set {_tests::2} to "clear message"
+	set {_tests::3} to "message test complete"
+	ensure junit test "org.skriptlang.skript.test.tests.syntaxes.expressions.ExprMessageTest" completes {_tests::*}
+
+on chat:
+	set {_test} to "org.skriptlang.skript.test.tests.syntaxes.expressions.ExprMessageTest"
+    junit test is {_test}
+
+    set chat message to "hello"
+    if message is "hello":
+    	complete objective "set message" for {_test}
+
+    clear message
+    if message is not set:
+    	complete objective "clear message" for {_test}
+
+    complete objective "message test complete" for {_test}
